Separate async task and cancellable
