+++ src/main/kotlin/com/projectcitybuild/core/contracts/PlayerStoreWrapper.kt
@@ -1,6 +1,6 @@
 package com.projectcitybuild.core.contracts
 
-import com.projectcitybuild.core.services.PlayerStore
+import com.projectcitybuild.core.utilities.PlayerStore
 
 interface PlayerStoreWrapper {
     val store: PlayerStore
