+++ src/main/kotlin/com/projectcitybuild/core/utilities/Cancellable.kt
@@ -1,0 +1,26 @@
+package com.projectcitybuild.core.utilities
+
+import java.util.concurrent.locks.ReentrantLock
+
+/**
+ * Thread-safe wrapper for anything that can be cancelled
+ */
+class Cancellable(private val handler: () -> Void) {
+    private val lock = ReentrantLock()
+
+    var isCancelled: Boolean = false
+        private set
+
+    fun cancel() {
+        lock.lock()
+        try {
+            if (isCancelled) return
+
+            handler()
+            isCancelled = true
+        }
+        finally {
+            lock.unlock()
+        }
+    }
+}
